requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
djangorestframework==3.14.0
future==0.18.2
psycopg2-binary==2.9.5
pytz==2022.6
serializers==0.2.4
sqlparse==0.4.3
